+++ nemo/collections/nlp/data/tokenizers/__init__.py
@@ -18,5 +18,6 @@ from nemo.collections.nlp.data.tokenizers.bert_tokenizer import NemoBertTokenize
 from nemo.collections.nlp.data.tokenizers.char_tokenizer import CharTokenizer
 from nemo.collections.nlp.data.tokenizers.gpt2_tokenizer import NemoGPT2Tokenizer
 from nemo.collections.nlp.data.tokenizers.sentencepiece_tokenizer import SentencePieceTokenizer
+from nemo.collections.nlp.data.tokenizers.utils import *
 from nemo.collections.nlp.data.tokenizers.word_tokenizer import WordTokenizer
 from nemo.collections.nlp.data.tokenizers.youtokentome_tokenizer import YouTokenToMeTokenizer
